Create lab2.asm
Wrote an MSP430 assembly language program where SORT1 sets the parameters for R4/R5/R6, which are used by the COPY and SORT subroutines to copy and sort array ARY1. ARY1 and ARY2 are the original un-altered arrays, while ARY1S and ARY2S are altered. R4 holds the starting address of the array. R5 holds the length of the array. R6 holds the starting location of the sorted array. The COPY subroutine copies the contents of array ARY1 into ARY1S. SORT subroutine sorts the elements on ARY1S in place. SORT2 is similar to SORT1 using same registers. An outline was provided.