Update lab3.asm
Wrote an MSP430 assembly language program to calculate the following arithmetic function: Screenshot 2023-04-17 at 6 37 42 PM The value of "a" is stored in register R4 as 5. The value of "X" is stored in R5. The value of "F" is stored in R7. An outline was provided.

Lab 4: Wrote a program in C and assembled a breadboard to create a 3-bit lighting pattern. A total of 6 switches controlled the output of 8 LED's on the breadboard. The first 3 switches would change the direction that lights activated and deactivated, as well as the speed that the light pattern rotated. The next 3 switches determined if the first, second, or third light in the pattern would be activated or deactivated. https://user-images.githubusercontent.com/112228115/232627692-7eaa20eb-f142-473f-ac87-a16f273979be.mov