Create lab4.asm
Wrote a program in C and assembled a breadboard to create a 3-bit lighting pattern. A total of 6 switches controlled the output of 8 LED's on the breadboard. The first 3 switches would change the direction that lights activated and deactivated, as well as the speed that the light pattern rotated. The next 3 switches determined if the first, second, or third light in the pattern would be activated or deactivated. https://user-images.githubusercontent.com/112228115/232627692-7eaa20eb-f142-473f-ac87-a16f273979be.mov